First pass at inherited TCV
